refactor: move setCameraPosition
