+++ src/main/java/org/runejs/client/Class49.java
@@ -12,35 +12,7 @@ public class Class49 {
     public static CacheArchive definitionCache;
     public static int anInt1154;
 
-    public static void setCameraPosition(int pitch, int x, int z, int yaw, int y, int ___pitch, int arg6) {
-        int xOffset = 0;
-        int yawDifference = 0x7ff & -yaw + 2048;
-        int zOffset = 0;
-        int pitchDifference = 2048 - pitch & 0x7ff;
-        int yOffset = arg6;
-        if(!Player.cutsceneActive) {
-            yOffset = SceneCamera.cameraZoom + pitch * 3;
-        }
-        if(pitchDifference != 0) {
-            int cosine = Model.COSINE[pitchDifference];
-            int sine = Model.SINE[pitchDifference];
-            int temp = zOffset * cosine + -(sine * yOffset) >> 16;
-            yOffset = cosine * yOffset + sine * zOffset >> 16;
-            zOffset = temp;
-        }
-        if(yawDifference != 0) {
-            int cosine = Model.COSINE[yawDifference];
-            int sine = Model.SINE[yawDifference];
-            int temp = cosine * xOffset + yOffset * sine >> 16;
-            yOffset = -(xOffset * sine) + yOffset * cosine >> 16;
-            xOffset = temp;
-        }
-        SceneCamera.cameraHorizontalRotation = yaw;
-        SceneCamera.cameraZ = -zOffset + z;
-        SceneCamera.cameraX = -xOffset + x;
-        SceneCamera.cameraVerticalRotation = pitch;
-        SceneCamera.cameraY = y + -yOffset;
-    }
+
 
 
 }
