+++ src/main/java/org/runejs/client/scene/SceneCamera.java
@@ -1,5 +1,8 @@
 package org.runejs.client.scene;
 
+import org.runejs.client.media.renderable.Model;
+import org.runejs.client.media.renderable.actor.Player;
+
 /**
  * Temporary holding ground for camera code
  *
@@ -29,4 +32,34 @@ public class SceneCamera {
     public static int[] customCameraTimer = new int[5];
     public static int[] customCameraFrequency = new int[5];
     public static int[] customCameraAmplitude = new int[5];
+
+    public static void setCameraPosition(int pitch, int x, int z, int yaw, int y, int ___pitch, int arg6) {
+        int xOffset = 0;
+        int yawDifference = 0x7ff & -yaw + 2048;
+        int zOffset = 0;
+        int pitchDifference = 2048 - pitch & 0x7ff;
+        int yOffset = arg6;
+        if(!Player.cutsceneActive) {
+            yOffset = SceneCamera.cameraZoom + pitch * 3;
+        }
+        if(pitchDifference != 0) {
+            int cosine = Model.COSINE[pitchDifference];
+            int sine = Model.SINE[pitchDifference];
+            int temp = zOffset * cosine + -(sine * yOffset) >> 16;
+            yOffset = cosine * yOffset + sine * zOffset >> 16;
+            zOffset = temp;
+        }
+        if(yawDifference != 0) {
+            int cosine = Model.COSINE[yawDifference];
+            int sine = Model.SINE[yawDifference];
+            int temp = cosine * xOffset + yOffset * sine >> 16;
+            yOffset = -(xOffset * sine) + yOffset * cosine >> 16;
+            xOffset = temp;
+        }
+        SceneCamera.cameraHorizontalRotation = yaw;
+        SceneCamera.cameraZ = -zOffset + z;
+        SceneCamera.cameraX = -xOffset + x;
+        SceneCamera.cameraVerticalRotation = pitch;
+        SceneCamera.cameraY = y + -yOffset;
+    }
 }
